feat!: remove ComposableProgram type
