--- packages/programs/program/CHANGELOG.md
@@ -1,176 +1,0 @@
-# Changelog
-
-### Dependencies
-
-* The following workspace dependencies were updated
-  * dependencies
-    * @peerbit/test-utils bumped from 1.0.0-alpha1 to 1.0.1-alpha1
-
-### Dependencies
-
-* The following workspace dependencies were updated
-  * dependencies
-    * @peerbit/pubsub bumped from 1.0.3 to 1.0.4
-    * @peerbit/crypto bumped from 1.0.3 to 1.0.4
-    * peerbit-libp2p bumped from 1.0.3 to 1.0.4
-    * @peerbit/log bumped from 1.0.3 to 1.0.4
-    * @peerbit/test-utils bumped from 1.0.3 to 1.0.4
-
-### Dependencies
-
-* The following workspace dependencies were updated
-  * dependencies
-    * @peerbit/crypto bumped from 1.0.0 to 1.0.1
-    * @peerbit/blocks-interface bumped from ^1.0.0 to ^1.0.1
-    * @peerbit/pubsub-interface bumped from ^1.0.0 to ^1.0.1
-
-### Dependencies
-
-* The following workspace dependencies were updated
-  * dependencies
-    * @peerbit/crypto bumped from 1.0.1 to 1.0.2
-    * @peerbit/blocks-interface bumped from ^1.0.1 to ^1.0.2
-    * @peerbit/pubsub-interface bumped from ^1.0.1 to ^1.0.2
-
-## [1.0.6](https://github.com/dao-xyz/peerbit/compare/program-v1.0.5...program-v1.0.6) (2023-07-04)
-
-
-### Bug Fixes
-
-* rm postbuild script ([b627bf0](https://github.com/dao-xyz/peerbit/commit/b627bf0dcdb99d24ac8c9055586e72ea2d174fcc))
-
-
-### Dependencies
-
-* The following workspace dependencies were updated
-  * dependencies
-    * @peerbit/crypto bumped from 1.0.2 to 1.0.3
-    * @peerbit/blocks-interface bumped from ^1.0.3 to ^1.0.4
-    * @peerbit/pubsub-interface bumped from ^1.0.3 to ^1.0.4
-    * @peerbit/lazy-level bumped from ^1.0.0 to ^1.0.1
-
-## [1.0.5](https://github.com/dao-xyz/peerbit/compare/program-v1.0.4...program-v1.0.5) (2023-07-03)
-
-
-### Bug Fixes
-
-* make drop conditional if only no other dependencies exist ([ec2407e](https://github.com/dao-xyz/peerbit/commit/ec2407e4b4f89a8e268004094a5dcb008de7d9e2))
-* remove redundant test ([4fb24fa](https://github.com/dao-xyz/peerbit/commit/4fb24fab54eeff94ca4b081dd80ae5065d075eab))
-
-## [1.0.4](https://github.com/dao-xyz/peerbit/compare/program-v1.0.3...program-v1.0.4) (2023-07-02)
-
-
-### Bug Fixes
-
-* add onBeforeCallback ([1977eee](https://github.com/dao-xyz/peerbit/commit/1977eee8c0629fd3a9931a78c271009b40040426))
-
-## [1.0.3](https://github.com/dao-xyz/peerbit/compare/program-v1.0.2...program-v1.0.3) (2023-06-29)
-
-
-### Bug Fixes
-
-* remove setMaxListeners call ([7331dd2](https://github.com/dao-xyz/peerbit/commit/7331dd26ec6c9d31229f30ca2d7524d78c79cb5f))
-
-
-### Dependencies
-
-* The following workspace dependencies were updated
-  * dependencies
-    * @peerbit/blocks-interface bumped from ^1.0.2 to ^1.0.3
-    * @peerbit/pubsub-interface bumped from ^1.0.2 to ^1.0.3
-
-## 1.0.0 (2023-06-28)
-
-
-### ⚠ BREAKING CHANGES
-
-* rename org on utility modules
-* reuse pubsub message id on rpc messages
-* client abstraction
-
-### Features
-
-* client abstraction ([6a1226d](https://github.com/dao-xyz/peerbit/commit/6a1226d4f8fc6deb167bff86cf7bdd6227c01a6b))
-* reuse pubsub message id on rpc messages ([57bede7](https://github.com/dao-xyz/peerbit/commit/57bede71cd822c71b439bd8011b6f25bff1da5cb))
-
-
-### Bug Fixes
-
-* rename org on utility modules ([0e09c8a](https://github.com/dao-xyz/peerbit/commit/0e09c8a29487205e02e45cc7f1e214450f96cb38))
-
-
-### Dependencies
-
-* The following workspace dependencies were updated
-  * dependencies
-    * @peerbit/crypto bumped from 1.0.4 to 1.0.0
-    * @peerbit/lazy-level bumped from ^0.0.9 to ^1.0.0
-
-## [1.0.3](https://github.com/dao-xyz/peerbit/compare/peerbit-program-v1.0.1-alpha1...peerbit-program-v1.0.3) (2023-06-14)
-
-
-### Bug Fixes
-
-* update invalid versions from prerelease release-please ([e2f6411](https://github.com/dao-xyz/peerbit/commit/e2f6411d46edf6d36723ca1ea81d1e55a09d3cd4))
-
-
-### Dependencies
-
-* The following workspace dependencies were updated
-  * dependencies
-    * @peerbit/pubsub bumped from 1.0.2 to 1.0.3
-    * @peerbit/crypto bumped from 1.0.2 to 1.0.3
-    * peerbit-libp2p bumped from 1.0.2 to 1.0.3
-    * @peerbit/log bumped from 1.0.2 to 1.0.3
-    * @peerbit/test-utils bumped from 1.0.2 to 1.0.3
-
-## 1.0.0-alpha1 (2023-06-14)
-
-
-### ⚠ BREAKING CHANGES
-
-* program identity derived from libp2p
-* throw error if variant is unspecified
-* simplify role names
-* simplified identity handling
-
-### Features
-
-* program identity derived from libp2p ([e7802f8](https://github.com/dao-xyz/peerbit/commit/e7802f816eb3e06c14cc57b193d2bde2b5005cef))
-* simplified identity handling ([1ae2416](https://github.com/dao-xyz/peerbit/commit/1ae24168a5c8629b8f9d1c57eceed6abd4a15020))
-* throw error if variant is unspecified ([f4aef0e](https://github.com/dao-xyz/peerbit/commit/f4aef0ea5713eb37a0dfcf251fe6233e6a54dbd7))
-
-
-### Bug Fixes
-
-* simplify role names ([f2bfd65](https://github.com/dao-xyz/peerbit/commit/f2bfd65422d0d7066cbc34693bfeafecb508004d))
-
-
-### Dependencies
-
-* The following workspace dependencies were updated
-  * dependencies
-    * @peerbit/pubsub bumped from 0.5.3 to 1.0.0-alpha1
-    * @peerbit/crypto bumped from 0.1.16 to 1.0.0-alpha1
-    * peerbit-libp2p bumped from 0.5.3 to 1.0.0-alpha1
-    * @peerbit/log bumped from 0.8.6 to 1.0.0-alpha1
-    * @peerbit/test-utils bumped from 0.4.3 to 1.0.0-alpha1
-
-## @peerbit/program [0.9.1](https://github.com/dao-xyz/peerbit/compare/@peerbit/program@0.9.0...@peerbit/program@0.9.1) (2023-06-07)
-
-
-### Bug Fixes
-
-* add release cfg ([de76654](https://github.com/dao-xyz/peerbit/commit/de766548f8106804d319e8b51e9607f2a3f60726))
-
-
-
-
-
-### Dependencies
-
-* **@peerbit/pubsub:** upgraded to 0.5.3
-* **@peerbit/crypto:** upgraded to 0.1.16
-* **peerbit-libp2p:** upgraded to 0.5.3
-* **@peerbit/log:** upgraded to 0.8.6
-* **@peerbit/test-utils:** upgraded to 0.4.3
